Add exact Decimal type for currency amounts (#89)
<!-- CURSOR_AGENT_PR_BODY_BEGIN -->
Adds `Decimal`, a Kotlin Multiplatform exact decimal for mobility
**currency** amounts (GBFS, GOFS, and GTFS Schedule). Coordinates and
distances stay as `Double`.

## Why
Issue #24: `Double` is a poor fit for prices and fares. This type stores
nine fractional digits in a `Long` and never goes through `Double`
except for explicit `toDouble()` / `toFloat()`.

## Design
- `@JvmInline value class Decimal` with factories `of(Int/Long)`,
`parse` / `parseOrNull`, and `ZERO` / `ONE` / `TEN`.
- Scale is always 9. Arithmetic is exact at that precision; `divide` and
`round` take `RoundingMode`.
- JSON uses `DecimalSerializer` (`JsonUnquotedLiteral` on encode).
Decode is strict by default (JSON numbers only). When the host `Json` is
lenient (`GbfsJson` / `GofsJson`), quoted numeric strings are accepted,
matching previous `Double` + lenient behavior.
- Parse and JSON both validate the **effective** scale after applying
any exponent. Digit collection is not limited to 18 digits, so values
like `9223372036854775807e-9` and `92233720368547758070e-10` parse when
they fit the final scaled Long.
- `CurrencyCode.defaultFractionDigits` uses ISO 4217 List One (SIX,
2026-01-01).
- GTFS `FareProduct` encodes `amount,currency` at the ISO 4217
minor-unit scale (`2.50,USD`; `250,JPY`). Decode still accepts extra
trailing zeros.

## Out of scope
- No `of(Double)` / `of(Float)`
- No GBFS/GOFS/GTFS Realtime coordinate or distance changes

Closes #24

<!-- CURSOR_AGENT_PR_BODY_END -->

<div><a
href="https://cursor.com/agents/bc-533855ee-6499-4f83-9556-02c13b5d2ba4?cursor_ref=pr_footer&cursor_cta=open_in_web"><picture><source
media="(prefers-color-scheme: dark)"
srcset="https://cursor.com/assets/images/open-in-web-dark.png"><source
media="(prefers-color-scheme: light)"
srcset="https://cursor.com/assets/images/open-in-web-light.png"><img
alt="Open in Web" width="114" height="28"
src="https://cursor.com/assets/images/open-in-web-dark.png"></picture></a>&nbsp;<a
href="https://cursor.com/background-agent?bcId=bc-533855ee-6499-4f83-9556-02c13b5d2ba4&cursor_ref=pr_footer&cursor_cta=open_in_cursor"><picture><source
media="(prefers-color-scheme: dark)"
srcset="https://cursor.com/assets/images/open-in-cursor-dark.png"><source
media="(prefers-color-scheme: light)"
srcset="https://cursor.com/assets/images/open-in-cursor-light.png"><img
alt="Open in Cursor" width="131" height="28"
src="https://cursor.com/assets/images/open-in-cursor-dark.png"></picture></a>&nbsp;</div>

---------

+++ gbfs-v1/src/commonMain/kotlin/dev/sargunv/mobilitydata/gbfs/v1/SystemPricingPlans.kt
@@ -1,6 +1,7 @@
 package dev.sargunv.mobilitydata.gbfs.v1
 
 import dev.sargunv.mobilitydata.utils.CurrencyCode
+import dev.sargunv.mobilitydata.utils.Decimal
 import dev.sargunv.mobilitydata.utils.IntBoolean
 import dev.sargunv.mobilitydata.utils.Url
 import kotlinx.serialization.SerialName
@@ -34,7 +35,7 @@ public data class PricingPlan(
   public val currency: CurrencyCode,
 
   /** Fare price, in the unit specified by currency. */
-  public val price: Double,
+  public val price: Decimal,
 
   /**
    * Will additional tax be added to the base price?
